+++ src/TidalControllers/MediaSessionController/MediaSessionController.ts
@@ -1,9 +1,9 @@
 import { Logger } from "../../features/logger";
 import { getTrackURL } from "../../features/tidal/url";
-import { convertDurationToSeconds } from "../../features/time/parse";
+import { convertSecondsToClockFormat } from "../../features/time/parse";
 import type { MediaInfo } from "../../models/mediaInfo";
 import { MediaStatus } from "../../models/mediaStatus";
-import type { RepeatState } from "../../models/repeatState";
+import type { RepeatStateType } from "../../models/repeatState";
 import { constrainPollingInterval } from "../../utility/pollingConstraints";
 import type { DomControllerOptions } from "../DomController/DomControllerOptions";
 import { DomTidalController } from "../DomController/DomTidalController";
@@ -62,6 +62,7 @@ export class MediaSessionController implements TidalController<MediaSessionContr
         if (mediaMetadata) {
           const current = this.getCurrentTime();
           const duration = this.getDuration();
+
           // Convert MediaMetadata to our MediaInfo format using internal getters
           const currentMediaInfo: Partial<MediaInfo> = {
             title: this.getTitle(),
@@ -71,13 +72,15 @@ export class MediaSessionController implements TidalController<MediaSessionContr
             playingFrom: this.getPlayingFrom(),
             status: this.getCurrentlyPlayingStatus(),
             url: getTrackURL(this.getTrackId()),
-            current,
-            currentInSeconds: convertDurationToSeconds(current),
-            duration,
-            durationInSeconds: convertDurationToSeconds(duration),
+            current: convertSecondsToClockFormat(current),
+            currentInSeconds: current,
+            duration: convertSecondsToClockFormat(duration),
+            durationInSeconds: duration,
             image: this.getSongImage(),
             icon: this.getSongIcon(),
             favorite: this.isFavorite(),
+            trackId: this.getTrackId(),
+            volume: this.getVolume(),
             player: {
               status: this.getCurrentlyPlayingStatus(),
               shuffle: this.getCurrentShuffleState(),
@@ -268,28 +271,32 @@ export class MediaSessionController implements TidalController<MediaSessionContr
     return this.fallbackDomController.getCurrentShuffleState();
   }
 
-  getCurrentRepeatState(): RepeatState {
+  getCurrentRepeatState(): RepeatStateType {
     return this.fallbackDomController.getCurrentRepeatState();
-  }
-
-  getCurrentPosition(): string {
-    return this.fallbackDomController.getCurrentPosition();
-  }
-
-  getCurrentPositionInSeconds(): number {
-    return this.fallbackDomController.getCurrentPositionInSeconds();
   }
 
   getTrackId(): string {
     return this.fallbackDomController.getTrackId();
   }
 
-  getCurrentTime(): string {
+  getCurrentTime(): number {
     return this.fallbackDomController.getCurrentTime();
   }
 
-  getDuration(): string {
+  setCurrentTime(time: number): void {
+    this.fallbackDomController.setCurrentTime(time);
+  }
+
+  getDuration(): number {
     return this.fallbackDomController.getDuration();
+  }
+
+  getVolume(): number {
+    return this.fallbackDomController.getVolume();
+  }
+
+  setVolume(volume: number) {
+    this.fallbackDomController.setVolume(volume);
   }
 
   getPlayingFrom(): string {