+++ src/constants/globalEvents.ts
@@ -16,4 +16,8 @@ export const globalEvents = {
   toggleShuffle: "toggleShuffle",
   toggleRepeat: "toggleRepeat",
   resetZoom: "resetZoom",
+  volume: "volume",
+  seek: "seek",
+  setLoopState: "setLoopState",
+  quit: "quit",
 };